pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bicameris"
version = "4.0.0"
description = "Kernel cognitif bicaméral et moteur d'arbitrage entropique - Hope 'n Mind"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "CC-BY-SA-4.0"}
authors = [
    {name = "Hope 'n Mind", email = "contact@hopemind.local"}
]
keywords = ["ai", "llm", "autonomous", "cognitive", "bipolar", "bicameral"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: CC-BY-SA-4.0 License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.6.0",
    "pynvml>=11.5.0",
    "psutil>=5.9.8",
    "llama-cpp-python>=0.2.56",
    "pyzmq>=25.1.2",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
full = [
    "qdrant-client>=1.17.0",
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.4.0",
    "beautifulsoup4>=4.12.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.3.0",
]

[project.scripts]
bicameris = "run:main"

[project.urls]
Homepage = "https://www.hopenmind.com"
Documentation = "https://github.com/hopemind/bicameris/docs"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "server*", "web*", "storage*"]
exclude = ["tests*", "sandbox*"]

[tool.setuptools.package-data]
core = ["**/*.py"]
server = ["**/*.py"]
web = ["templates/**/*", "static/**/*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
